Add InspectFromClick script for object inspection functionality
Added `InspectFromClick` that allows objects to be inspected by clicking. The script manages the camera and post-processing effects during inspection mode, ensuring a focused view on the selected object while deactivating other elements. This lets us see the plants up close in the card game environment.

Disabled the automatic application of selected cards on plants with 'Click3D' and 'PlantCardFunctions'.

Signed-off-by: donovan <dmontoya@unity.edu>